Various bug fixes
