+++ app/src/main/java/com/nextcloud/talk/newarch/di/module/NetworkModule.kt
@@ -156,7 +156,11 @@ fun createOkHttpClient(
         var response = chain.proceed(chain.request())
 
         if (response.request().url().encodedPath().contains("/avatar/")) {
-            AvatarStatusCodeHolder.getInstance().statusCode = response.code()
+            if (response.header("x-nc-iscustomavatar", "0") == "1") {
+                AvatarStatusCodeHolder.getInstance().statusCode = response.code()
+            } else  {
+                AvatarStatusCodeHolder.getInstance().statusCode = response.code()
+            }
 
             if (response.code() == 201) {
                 response = response.newBuilder().code(200).message("OK").build()
