+++ app/src/main/java/com/nextcloud/talk/newarch/services/CallService.kt
@@ -76,8 +76,17 @@ class CallService : Service(), KoinComponent, CoroutineScope {
 
     override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
         intent?.let {
-            if (it.action == BundleKeys.KEY_INCOMING_PUSH_MESSSAGE) {
-                decryptMessage(it.getStringExtra(BundleKeys.KEY_ENCRYPTED_SUBJECT), it.getStringExtra(BundleKeys.KEY_ENCRYPTED_SIGNATURE))
+            if (it.action == BundleKeys.KEY_SHOW_INCOMING_CALL) {
+                val decryptedPushMessageString: String = it.getStringExtra(BundleKeys.KEY_DECRYPTED_PUSH_MESSAGE)!!
+                val signatureVerificationString: String = it.getStringExtra(BundleKeys.KEY_SIGNATURE_VERIFICATION)!!
+                val conversationString: String = it.getStringExtra(BundleKeys.KEY_CONVERSATION)!!
+
+                val json = Json(MagicJson.customJsonConfiguration)
+                val decryptedPushMessage = LoganSquare.parse(decryptedPushMessageString, DecryptedPushMessage::class.java)
+                val signatureVerification = json.parse(SignatureVerification.serializer(), signatureVerificationString)
+                val conversation = json.parse(Conversation.serializer(), conversationString)
+
+                showIncomingCall(signatureVerification, decryptedPushMessage, conversation)
             } else if (it.action == BundleKeys.KEY_REJECT_INCOMING_CALL || it.action == BundleKeys.DISMISS_CALL_NOTIFICATION) {
                 if (it.getStringExtra(BundleKeys.KEY_ACTIVE_NOTIFICATION) == activeNotification) {
                     endIncomingConversation(it.action != BundleKeys.DISMISS_CALL_NOTIFICATION)
@@ -92,133 +101,74 @@ class CallService : Service(), KoinComponent, CoroutineScope {
 
     }
 
-    private fun decryptMessage(subject: String, signature: String) = coroutineContext.run {
-        launch {
-            val signatureVerification: SignatureVerification
-            val decryptedPushMessage: DecryptedPushMessage
+    private fun showIncomingCall(signatureVerification: SignatureVerification, decryptedPushMessage: DecryptedPushMessage, conversation: Conversation) {
+        val generatedActiveNotificationId = signatureVerification.userEntity!!.id.toString() + "@" + decryptedPushMessage.notificationId!!.toString()
+        val fullScreenPendingIntent = NotificationUtils.getIncomingCallIPendingIntent(applicationContext, this@CallService, conversation, signatureVerification.userEntity!!, generatedActiveNotificationId)
 
-            try {
-                val base64DecodedSubject = Base64.decode(subject, Base64.DEFAULT)
-                val base64DecodedSignature = Base64.decode(signature, Base64.DEFAULT)
-                val pushUtils = PushUtils(usersRepository)
-                val privateKey = pushUtils.readKeyFromFile(false) as PrivateKey
-                try {
-                    signatureVerification = pushUtils.verifySignature(base64DecodedSignature, base64DecodedSubject)
-                    if (signatureVerification.signatureValid) {
-                        val cipher = Cipher.getInstance("RSA/None/PKCS1Padding")
-                        cipher.init(Cipher.DECRYPT_MODE, privateKey)
-                        val decryptedSubject = cipher.doFinal(base64DecodedSubject)
-                        decryptedPushMessage = LoganSquare.parse(String(decryptedSubject), DecryptedPushMessage::class.java)
-                        val conversation = getConversationForTokenAndUser(signatureVerification.userEntity!!, decryptedPushMessage.id!!)
+        val soundUri = NotificationUtils.getCallSoundUri(applicationContext, appPreferences)
+        val vibrationEffect = NotificationUtils.getVibrationEffect(appPreferences)
 
-                        decryptedPushMessage.apply {
-                            when {
-                                delete -> {
-                                    NotificationUtils.cancelExistingNotificationWithId(applicationContext, signatureVerification.userEntity!!, notificationId!!)
-                                }
-                                deleteAll -> {
-                                    NotificationUtils.cancelAllNotificationsForAccount(applicationContext, signatureVerification.userEntity!!)
-                                }
-                                type == "call" && (conversation?.type == Conversation.ConversationType.ONE_TO_ONE_CONVERSATION || conversation?.objectType.equals("share:password")) -> {
-                                    if (conversation != null) {
-                                        val generatedActiveNotificationId = signatureVerification.userEntity!!.id.toString() + "@" + decryptedPushMessage.notificationId!!.toString()
-                                        val fullScreenPendingIntent = NotificationUtils.getIncomingCallIPendingIntent(applicationContext, this@CallService, conversation, signatureVerification.userEntity!!, generatedActiveNotificationId)
+        val notificationChannelId = NotificationUtils.getNotificationChannelId(applicationContext, applicationContext.resources
+                .getString(R.string.nc_notification_channel_calls), applicationContext.resources
+                .getString(R.string.nc_notification_channel_calls_description), true,
+                NotificationManagerCompat.IMPORTANCE_HIGH, soundUri!!,
+                NotificationUtils.getCallAudioAttributes(true), vibrationEffect, false, null)
 
-                                        val soundUri = NotificationUtils.getCallSoundUri(applicationContext, appPreferences)
-                                        val vibrationEffect = NotificationUtils.getVibrationEffect(appPreferences)
+        val uri: Uri = Uri.parse(signatureVerification.userEntity?.baseUrl)
+        var baseUrl = uri.host
 
-                                        val notificationChannelId = NotificationUtils.getNotificationChannelId(applicationContext, applicationContext.resources
-                                                .getString(R.string.nc_notification_channel_calls), applicationContext.resources
-                                                .getString(R.string.nc_notification_channel_calls_description), true,
-                                                NotificationManagerCompat.IMPORTANCE_HIGH, soundUri!!,
-                                                NotificationUtils.getCallAudioAttributes(true), vibrationEffect, false, null)
+        if (baseUrl == null) {
+            baseUrl = signatureVerification.userEntity?.baseUrl
+        }
 
-                                        val uri: Uri = Uri.parse(signatureVerification.userEntity?.baseUrl)
-                                        var baseUrl = uri.host
+        var largeIcon = BitmapFactory.decodeResource(applicationContext.resources, R.drawable.ic_baseline_person_black_24)
 
-                                        if (baseUrl == null) {
-                                            baseUrl = signatureVerification.userEntity?.baseUrl
-                                        }
+        val rejectCallIntent = Intent(this@CallService, CallService::class.java)
+        rejectCallIntent.action = BundleKeys.KEY_REJECT_INCOMING_CALL
+        rejectCallIntent.putExtra(BundleKeys.KEY_ACTIVE_NOTIFICATION, generatedActiveNotificationId)
+        val rejectCallPendingIntent = PendingIntent.getService(this@CallService, 0, rejectCallIntent, PendingIntent.FLAG_UPDATE_CURRENT)
+        val notificationBuilder = NotificationCompat.Builder(this@CallService, notificationChannelId)
+                .setPriority(NotificationCompat.PRIORITY_HIGH)
+                .setCategory(NotificationCompat.CATEGORY_CALL)
+                .setSmallIcon(R.drawable.ic_call_black_24dp)
+                .setLargeIcon(largeIcon)
+                .setSubText(baseUrl)
+                .setShowWhen(true)
+                .setWhen(System.currentTimeMillis())
+                .setContentTitle(EmojiCompat.get().process(decryptedPushMessage.subject.toString()))
+                .setAutoCancel(true)
+                .setOngoing(true)
+                .addAction(R.drawable.ic_call_end_white_24px, resources.getString(R.string.nc_reject_call), rejectCallPendingIntent)
+                .setContentIntent(fullScreenPendingIntent)
+                .setFullScreenIntent(fullScreenPendingIntent, true)
+                .setSound(NotificationUtils.getCallSoundUri(applicationContext, appPreferences), AudioManager.STREAM_RING)
 
-                                        var largeIcon = BitmapFactory.decodeResource(applicationContext.resources, R.drawable.ic_baseline_person_black_24)
+        if (vibrationEffect != null) {
+            notificationBuilder.setVibrate(vibrationEffect)
+        }
 
-                                        val rejectCallIntent = Intent(this@CallService, CallService::class.java)
-                                        rejectCallIntent.action = BundleKeys.KEY_REJECT_INCOMING_CALL
-                                        rejectCallIntent.putExtra(BundleKeys.KEY_ACTIVE_NOTIFICATION, generatedActiveNotificationId)
-                                        val rejectCallPendingIntent = PendingIntent.getService(this@CallService, 0, rejectCallIntent, PendingIntent.FLAG_UPDATE_CURRENT)
-                                        val notificationBuilder = NotificationCompat.Builder(this@CallService, notificationChannelId)
-                                                .setPriority(NotificationCompat.PRIORITY_HIGH)
-                                                .setCategory(NotificationCompat.CATEGORY_CALL)
-                                                .setSmallIcon(R.drawable.ic_call_black_24dp)
-                                                .setLargeIcon(largeIcon)
-                                                .setSubText(baseUrl)
-                                                .setShowWhen(true)
-                                                .setWhen(System.currentTimeMillis())
-                                                .setContentTitle(EmojiCompat.get().process(decryptedPushMessage.subject.toString()))
-                                                .setAutoCancel(true)
-                                                .setOngoing(true)
-                                                .addAction(R.drawable.ic_call_end_white_24px, resources.getString(R.string.nc_reject_call), rejectCallPendingIntent)
-                                                .setContentIntent(fullScreenPendingIntent)
-                                                .setFullScreenIntent(fullScreenPendingIntent, true)
-                                                .setSound(NotificationUtils.getCallSoundUri(applicationContext, appPreferences), AudioManager.STREAM_RING)
+        val target = object : Target {
+            override fun onSuccess(result: Drawable) {
+                super.onSuccess(result)
+                largeIcon = result.toBitmap()
+                notificationBuilder.setLargeIcon(largeIcon)
+                showNotification(notificationBuilder, signatureVerification.userEntity!!, conversation!!.token!!, decryptedPushMessage.notificationId!!, generatedActiveNotificationId)
+            }
 
-                                        if (vibrationEffect != null) {
-                                            notificationBuilder.setVibrate(vibrationEffect)
-                                        }
-
-                                        val target = object : Target {
-                                            override fun onSuccess(result: Drawable) {
-                                                super.onSuccess(result)
-                                                largeIcon = result.toBitmap()
-                                                notificationBuilder.setLargeIcon(largeIcon)
-                                                showNotification(notificationBuilder, signatureVerification.userEntity!!, conversation.token!!, decryptedPushMessage.notificationId!!, generatedActiveNotificationId)
-                                            }
-
-                                            override fun onError(error: Drawable?) {
-                                                super.onError(error)
-                                                showNotification(notificationBuilder, signatureVerification.userEntity!!, conversation.token!!, decryptedPushMessage.notificationId!!, generatedActiveNotificationId)
-                                            }
-                                        }
-
-                                        val avatarUrl = ApiUtils.getUrlForAvatarWithName(signatureVerification.userEntity!!.baseUrl, conversation.name, R.dimen.avatar_size)
-                                        val imageLoader = networkComponents.getImageLoader(signatureVerification.userEntity!!.toUser())
-
-                                        val request = Images().getRequestForUrl(
-                                                imageLoader, applicationContext, avatarUrl, signatureVerification.userEntity!!.toUser(),
-                                                target, null, CircleCropTransformation())
-
-                                        imageLoader.load(request)
-                                    }
-                                }
-                                else -> {
-                                    if (conversation != null) {
-                                        val json = Json(MagicJson.customJsonConfiguration)
-
-                                        val messageData = Data.Builder()
-                                                .putString(BundleKeys.KEY_DECRYPTED_PUSH_MESSAGE, LoganSquare.serialize(decryptedPushMessage))
-                                                .putString(BundleKeys.KEY_SIGNATURE_VERIFICATION, json.stringify(SignatureVerification.serializer(), signatureVerification))
-                                                .putString(BundleKeys.KEY_CONVERSATION, json.stringify(Conversation.serializer(), conversation))
-                                                .build()
-                                        val pushNotificationWork = OneTimeWorkRequest.Builder(NotificationWorker::class.java).setInputData(messageData).build()
-                                        WorkManager.getInstance().enqueue(pushNotificationWork)
-                                    }
-                                }
-                            }
-                        }
-                    } else {
-                        // do absolutely nothing
-                    }
-                } catch (e1: NoSuchAlgorithmException) {
-                    Log.d(tag, "No proper algorithm to decrypt the message " + e1.localizedMessage)
-                } catch (e1: NoSuchPaddingException) {
-                    Log.d(tag, "No proper padding to decrypt the message " + e1.localizedMessage)
-                } catch (e1: InvalidKeyException) {
-                    Log.d(tag, "Invalid private key " + e1.localizedMessage)
-                }
-            } catch (exception: Exception) {
-                Log.d(tag, "Something went very wrong " + exception.localizedMessage)
+            override fun onError(error: Drawable?) {
+                super.onError(error)
+                showNotification(notificationBuilder, signatureVerification.userEntity!!, conversation!!.token!!, decryptedPushMessage.notificationId!!, generatedActiveNotificationId)
             }
         }
+
+        val avatarUrl = ApiUtils.getUrlForAvatarWithName(signatureVerification.userEntity!!.baseUrl, conversation!!.name, R.dimen.avatar_size)
+        val imageLoader = networkComponents.getImageLoader(signatureVerification.userEntity!!.toUser())
+
+        val request = Images().getRequestForUrl(
+                imageLoader, applicationContext, avatarUrl, signatureVerification.userEntity!!.toUser(),
+                target, null, CircleCropTransformation())
+
+        imageLoader.load(request)
     }
 
     private fun checkIsConversationActive(user: UserNgEntity, conversationToken: String, activeNotificationArgument: String) {
@@ -254,28 +204,6 @@ class CallService : Service(), KoinComponent, CoroutineScope {
         if (triggerEventBus) {
             eventBus.post(CallEvent())
         }
-    }
-
-    private suspend fun getConversationForTokenAndUser(user: UserNgEntity, conversationToken: String): Conversation? {
-        var conversation = conversationsRepository.getConversationForUserWithToken(user.id, conversationToken)
-        if (conversation == null) {
-            val getConversationUseCase = GetConversationUseCase(networkComponents.getRepository(false, user.toUser()), apiErrorHandler)
-            runBlocking {
-                getConversationUseCase.invoke(this, parametersOf(user, conversationToken), object : UseCaseResponse<ConversationOverall> {
-                    override suspend fun onSuccess(result: ConversationOverall) {
-                        val internalConversation = result.ocs.data
-                        conversationsRepository.saveConversationsForUser(user.id, listOf(internalConversation), false)
-                        conversation = result.ocs.data
-                    }
-
-                    override suspend fun onError(errorModel: ErrorModel?) {
-                        conversation = null
-                    }
-                })
-            }
-        }
-
-        return conversation
     }
 
     private fun showNotification(builder: NotificationCompat.Builder, user: UserNgEntity, conversationToken: String, internalNotificationId: Long, generatedNotificationId: String) {
